# Conflicts:
#	cpp/open3d/visualization/gui/SceneWidget.cpp

+++ cpp/open3d/visualization/gui/Events.h
@@ -66,8 +66,8 @@ struct MouseEvent {
             int count;
         } button;
         struct {
-            int dx;
-            int dy;
+            float dx;  // macOS gives fractional values, and is required
+            float dy;  //   for the buttery-smooth trackpad scrolling on macOS
             bool isTrackpad;
         } wheel;
     };
